+++ apps/client/src/App/App.test.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
 import { describe, it } from 'vitest';
-import App from './App';
+import App from '../app/App';
 
 describe('App rendered', () => {
   it('', () => {
